requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.6.0
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
et-xmlfile==1.1.0
faiss-cpu==1.13.2
fonttools==4.61.1
frozenlist==1.8.0
greenlet==3.3.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.3
idna==3.10
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.2.7
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-ollama==1.0.1
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.6
lark==1.2.2
marshmallow==3.26.2
matplotlib==3.10.8
multidict==6.7.1
mypy_extensions==1.1.0
neo4j==5.26.0
numpy==2.1.2
ollama==0.6.1
openai==2.16.0
openpyxl==3.1.5
orjson==3.11.5
ormsgpack==1.12.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
pillow==12.0.0
propcache==0.4.1
pycparser==2.23
pydantic==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.27.2
pyparsing==3.3.1
pypdf==5.0.1
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
selenium==4.39.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sortgs==1.0.7
soupsieve==2.8.1
SQLAlchemy==2.0.46
structured-logprobs==0.1.5
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.66.5
trio==0.32.0
trio-websocket==0.12.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
urllib3==2.6.2
uuid_utils==0.14.0
webdriver-manager==4.0.2
websocket-client==1.9.0
wsproto==1.3.2
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
